requirements.txt
streamlit
scikit-learn
pytorch
torchvision
PIL
joblib
pandas
matplotlib